+++ src/components/sdk/MeetingControls/AudioInputVFControl.tsx
@@ -1,6 +1,7 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
-import _ from 'lodash';
+
+import isEqual from 'lodash.isequal';
 
 import React, { useEffect, useState, ReactNode } from 'react';
 import { Device, AudioTransformDevice, VoiceFocusTransformDevice } from 'amazon-chime-sdk-js';
@@ -44,7 +45,7 @@ const AudioInputVFControl: React.FC<Props> = ({
   const meetingManager = useMeetingManager();
   const [isLoading, setIsLoading] = useState(false);
   // When the user click on Amazon Voice Focus option, the state will change.
-  const [isVoiceFocusOn, setIsVoiceFocusOn] = useState(false);
+  const [isVoiceFocusChecked, setIsVoiceFocusChecked] = useState(false);
   // Only when the current input audio device is an Amazon Voice Focus device, the state will be true. Otherwise, it will be false.
   const [isVoiceFocusEnabled, setIsVoiceFocusEnabled] = useState(false);
   const [dropdownWithVFOptions, setDropdownWithVFOptions] = useState<ReactNode[] | null>(null);
@@ -57,7 +58,7 @@ const AudioInputVFControl: React.FC<Props> = ({
   const { devices, selectedDevice } = useAudioInputs(audioInputConfig);
 
   const audioInputDevices: DeviceType[] = useMemoCompare(devices, (prev: DeviceType[], next: DeviceType[]): boolean => {
-    return _.isEqual(prev, next)
+    return isEqual(prev, next)
   });
 
   useEffect(() => {
@@ -66,6 +67,10 @@ const AudioInputVFControl: React.FC<Props> = ({
       meetingManager.unsubscribeFromSelectedAudioInputTransformDevice(setDevice);
     };
   }, []);
+
+  useEffect(() => {
+    console.info(`Amazon Voice Focus is ${isVoiceFocusEnabled ? 'enabled' : 'disabled'}.`);
+  }, [isVoiceFocusEnabled]);
 
   useEffect(() => {
     // Only when the current input audio device is an Amazon Voice Focus transform device,
@@ -84,7 +89,7 @@ const AudioInputVFControl: React.FC<Props> = ({
         children={<span>{device.label}</span>}
         checked={isOptionActive(selectedDevice, device.deviceId)}
         onClick={async (): Promise<void> => {
-          if (isVoiceFocusOn && !isLoading) {
+          if (isVoiceFocusChecked && !isLoading) {
             setIsLoading(true);
             const receivedDevice = audioInputSelectionToDevice(device.deviceId);
             const currentDevice = await addVoiceFocus(receivedDevice);
@@ -110,7 +115,7 @@ const AudioInputVFControl: React.FC<Props> = ({
           disabled={isLoading}
           onClick={() => {
             setIsLoading(true);
-            setIsVoiceFocusOn(current => !current);
+            setIsVoiceFocusChecked(current => !current);
           }} />
       )
       dropdownOptions?.push(<PopOverSeparator key = 'separator' />);
@@ -127,7 +132,7 @@ const AudioInputVFControl: React.FC<Props> = ({
     audioInputDevices,
     isLoading,
     isVoiceFocusEnabled,
-    isVoiceFocusOn,
+    isVoiceFocusChecked,
     isVoiceFocusSupported,
     selectedDevice,
   ]);
@@ -135,12 +140,14 @@ const AudioInputVFControl: React.FC<Props> = ({
   useEffect(() => {
     async function onVFCheckboxChange() {
       let current = device;
-      if (isVoiceFocusOn) {
+      if (isVoiceFocusChecked) {
+        console.info('User turned on Amazon Voice Focus.')
         if (typeof (device) === 'string') {
           const currentDevice = audioInputSelectionToDevice(device);
           current = await addVoiceFocus(currentDevice);
         }
       } else {
+        console.info('Amazon Voice Focus is off by default or user turned off Amazon Voice Focus.')
         if (device instanceof VoiceFocusTransformDevice) {
           current = device.getInnerDevice();
         }
@@ -150,7 +157,7 @@ const AudioInputVFControl: React.FC<Props> = ({
     }
 
     onVFCheckboxChange();
-  }, [isVoiceFocusOn]);
+  }, [isVoiceFocusChecked]);
 
   return (
     <ControlBarButton