+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Add logs for Amazon Voice Focus components.
+
 ### Changed
 
 - Update `useSelectVideoInputDevice` hook documentation and usage example.